D3DFMIQ-1819 Review related changes
Former-commit-id: 2435479c77c73f168804a6baa0e256fd2d3b6118
Former-commit-id: 3a8d2334f9a3ef1c12e17de0438916aca1d14ee8

+++ test/DeltaShell.Plugins.FMSuite.Wave.Tests/IO/Helpers/Boundaries/BoundaryCategoryConverterTest.cs
@@ -16,7 +16,7 @@ namespace DeltaShell.Plugins.FMSuite.Wave.Tests.IO.Helpers.Boundaries
         private readonly Random random = new Random();
 
         private double RandomDouble => Math.Round(random.NextDouble(), 3);
-        private double RandomInt => random.Next(100);
+        private int RandomInt => random.Next(100);
 
         [Test]
         public void Convert_BoundaryCategoryNull_ThrowsArgumentNullException()
